+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.63 fix: fix get_czce_rank_table interface
+
+    1. 修复 get_czce_rank_table 接口
+
 1.9.62 fix: fix bond_zh_us_rate interface
 
     1. 修复 bond_zh_us_rate 接口
@@ -2226,6 +2230,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.63 fix: fix get_czce_rank_table interface
 
 1.9.62 fix: fix bond_zh_us_rate interface